requirements.txt

Django==5.0.3
djangorestframework==3.15.1
drf-yasg==1.21.8   
django-cors-headers==4.3.1
